+++ catalog/ui/catalog-ui-search/src/test/java/org/codice/ddf/catalog/ui/forms/model/JsonTransformVisitorTest.java
@@ -20,17 +20,12 @@ import static org.codice.ddf.catalog.ui.forms.model.FilterNodeAssertionSupport.a
 
 import java.io.File;
 import java.io.FileInputStream;
-import java.io.InputStream;
 import java.net.URL;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Arrays;
-import javax.xml.bind.JAXBContext;
-import javax.xml.bind.JAXBElement;
-import javax.xml.bind.JAXBException;
-import javax.xml.bind.Unmarshaller;
-import net.opengis.filter.v_2_0.FilterType;
 import org.codice.ddf.catalog.ui.forms.SearchFormsLoaderTest;
+import org.codice.ddf.catalog.ui.forms.filter.FilterReader;
 import org.codice.ddf.catalog.ui.forms.filter.VisitableXmlElement;
 import org.codice.ddf.catalog.ui.forms.filter.VisitableXmlElementImpl;
 import org.junit.Before;
@@ -117,29 +112,6 @@ public class JsonTransformVisitorTest {
     }
 
     return new VisitableXmlElementImpl(
-        new FilterReader().unmarshal(new FileInputStream(xmlFile), FilterType.class));
-  }
-
-  private static class FilterReader {
-    private final JAXBContext context;
-
-    public FilterReader() throws JAXBException {
-      String pkgName = FilterType.class.getPackage().getName();
-      this.context = JAXBContext.newInstance(format("%s:%s", pkgName, pkgName));
-    }
-
-    @SuppressWarnings("unchecked")
-    public <T> JAXBElement<T> unmarshal(InputStream inputStream, Class<T> tClass)
-        throws JAXBException {
-      Unmarshaller unmarshaller = context.createUnmarshaller();
-      Object result = unmarshaller.unmarshal(inputStream);
-      if (result instanceof JAXBElement) {
-        JAXBElement element = (JAXBElement) result;
-        if (tClass.isInstance(element.getValue())) {
-          return (JAXBElement<T>) element;
-        }
-      }
-      return null;
-    }
+        new FilterReader().unmarshalFilter(new FileInputStream(xmlFile)));
   }
 }
